Who is looking is a leaf, not a serving idea
identityOf is the person. GET /olai/who is the door. Putting the parse
in @olai/server would make a later POST /capture import the composition
root. @olai/identity has no Effect, no HTTP, no @olai/* sibling — the
same cut @olai/git and @olai/fonts already are. The picture (gravatar)
is a second fold in the same package, not a hash mixed into the parse.
