+++ packages/server/README.md
@@ -32,7 +32,7 @@ Talking to the agent is not here. It was, and it was four modules of domain insi
 | `context.ts` | what the composer was ARMED with, resolved: the browser sends the ids a message is about and this reads them against the same snapshot a keystroke's write is judged against — title, `file:line`, ancestors — so the agent is told what the node IS rather than what a tab drew a frame ago. An id the set no longer declares refuses the send, in `@olai/ops`' own `notFound` words: a message about a node, sent without the node, is one whose subject the agent would have to guess. Pure over a reading, like its neighbour |
 | `published.ts` | the projection that fiber publishes: one published revision cut into per-file entries — an outline's nodes, a document's body, and the HEAD beside it, which is that same file's key carrying its revision and nothing else — with the store's own `changed`/`removed` mapped onto each collection's upserts and removes. One function over one binding of one list, because the two members keyed by a bodied file promise the same key set — the promise a reader's file list rests on |
 | `bodies.ts` | the bodies nobody keeps: the set holds a `.html`'s PATH and not its bytes (`@olai/format`'s `kinds.ts`), so a per-key `get` — the wire's own word for "somebody is showing this file" — reads it, publishes it on that key and lets it go. A revision that moved a file somebody is HOLDING re-reads it for them; one that moved four hundred saved pages nobody has open opens none of them. Who is holding what is the wire's own subscription lifetime, which the collection reports through its own `holders` dep (`./runtime.ts`), counted per path — first reader in starts the re-reads, last one out ends them — so there is no capacity and nothing to evict. What it holds is a count per path, one read in flight, and never a body. The browser is no longer one of those readers: a preview asks for the file's HEAD instead, so a saved page is read here for the agent that asks for its bytes and for nobody else. A read that is REFUSED publishes `DocumentEntry.refused` rather than nothing, so a one-shot reader is handed a frame and the page can say so |
-| `identity.ts` | who this REQUEST is: one configurable pair of trusted header names (login + optional email; default `Tailscale-User-Login` for both), `identityOf` the one reading, `GET /olai/who` the chip's door. The header is only meaningful when the proxy is the only way in and strips client-supplied copies. Not a cell — cells are one value for the process. The gravatar origin the chip fetches is admitted on the app page's own content policy, in the shell |
+| `identity.ts` | `GET /olai/who`, the chip's door: maps `@olai/identity`'s person onto the surface's `Who` and serves it. The reading itself is not here — putting it here would make a later `POST /capture` import this package for a parse. The gravatar origin the chip fetches is admitted on the app page's own content policy, in the shell |
 | `listener.ts` | one `serveSurfaceApp` call, and the one decision it leaves that is a policy: whose port this is |
 | `report.ts` | the other one: how surface-layer trouble READS in olai's log — a stale tab, a refused origin, a faulted connection, and a live stream that could not re-read its answer. Two callers (the listener's narration arm and `runtime.ts`'s stream deps) and one voice, so the levels and the annotations are compared in one file rather than across two |
 | `media.ts` | `/media/*`: the pictures a document points at and the pages a preview shows, sealed — the only bytes that leave the served directory over HTTP without going through the store |
