+++ packages/server/package.json
@@ -12,12 +12,13 @@
     "typecheck": "tsc --noEmit"
   },
   "//main": "src/serve.ts, not src/main.ts: main.ts ends in a top-level run that starts the server and installs signal handlers, so a resolver honouring `main` over `exports` would boot a listener on import. The binary is reached as a script path (default.nix, the justfile), never as a module.",
-  "//dependencies": "The composition root: it is the only package allowed to know about all of the others, and the direction is strictly downward (docs/architecture.md). @olai/web is deliberately absent — the server serves a built bundle it is handed, it does not import the client.",
+  "//dependencies": "The composition root: it is the only package allowed to know about all of the others, and the direction is strictly downward (docs/architecture.md). @olai/identity is the person on the request, as a value — this package is the HTTP door over it, not the parse. @olai/web is deliberately absent — the server serves a built bundle it is handed, it does not import the client.",
   "//acp": "The ACP SDK is NOT here: talking to an agent is @olai/chat's, and this package composes it. What is left of the agent in this file is a workspace dependency and one call.",
   "dependencies": {
     "@effect/platform-node": "4.0.0-rc.110",
     "@olai/chat": "workspace:*",
     "@olai/format": "workspace:*",
+    "@olai/identity": "workspace:*",
     "@olai/log": "workspace:*",
     "@olai/ops": "workspace:*",
     "@olai/store": "workspace:*",
