+++ packages/server/src/serve.ts
@@ -24,6 +24,7 @@
  */
 
 import { adapterFrom, AGENT_ENV, whyNoAgent } from "@olai/chat"
+import type { IdentityHeaders } from "@olai/identity"
 import { type CommitMode, make as makeOps, TOOLS } from "@olai/ops"
 import { Effect, SubscriptionRef } from "effect"
 import { randomBytes } from "node:crypto"
@@ -33,7 +34,6 @@ import * as path from "node:path"
 import * as Chat from "@olai/chat"
 import { openDirectory } from "./directory.ts"
 import { watchFault } from "./fault.ts"
-import type { IdentityHeaders } from "./identity.ts"
 import { listen } from "./listener.ts"
 import { clientOver, serveFace } from "./mcp/face.ts"
 import { MCP_PATH, mcpTransport } from "./mcp/route.ts"
@@ -51,7 +51,7 @@ export interface ServeOptions {
   /** Browser origins allowed to open the websocket, beyond same-origin. */
   readonly allowedOrigins: ReadonlyArray<string>
   /** Which request headers name the signed-in person — see
-   *  {@link ./identity.ts}. */
+   *  `@olai/identity`. */
   readonly identity: IdentityHeaders
   /** How writes reach git — `--commit=off | manual | auto`, `manual` by
    *  default. See `@olai/ops`'s `Options`. */
