+++ packages/surface/README.md
@@ -114,7 +114,7 @@ Two procedures go the other way and are the BROWSER's alone, and neither is a re
 
 PROCESS identity is deliberately NOT a member here. It is a real question — a page bound to a server that has been replaced must know, and both ends of the stale-tab handshake compare that id — but the framework reserves `system/identity` for it and answers it out of every surface, process id included. A member of our own would be a second answer to a question already answered.
 
-**Who is looking is a different question, and is not a member either.** A trusted login header on this request (default `Tailscale-User-Login`; the pair is configurable), or nobody when it is absent. The **path** and the JSON it carries are `src/who.ts` — the same bijection `/media/…` is, because the chip fetches and the server answers from packages that cannot import each other. The **reading** of the headers is `@olai/server`'s `identityOf`. The websocket cannot see it today — kolu's `serveSurfaceApp` owns the upgrade and does not hand request headers to the app.
+**Who is looking is a different question, and is not a member either.** A trusted login header on this request (default `Tailscale-User-Login`; the pair is configurable), or nobody when it is absent. The **path** and the JSON it carries are `src/who.ts` — the same bijection `/media/…` is, because the chip fetches and the server answers from packages that cannot import each other. The **reading** of the headers is `@olai/identity`'s `identityOf`. The websocket cannot see it today — kolu's `serveSurfaceApp` owns the upgrade and does not hand request headers to the app.
 
 ## One FACT that is not a member — and no longer here
