+++ packages/surface/src/index.ts
@@ -71,7 +71,7 @@
  * surface, process id included, so an app that declares its own is declaring a
  * second answer to a question already answered (juspay/kolu#2133). Who is
  * LOOKING is a different question and is not a member either: it is per HTTP
- * request (`@olai/server`'s `identity.ts`, `GET /olai/who`), because a cell
+ * request (`@olai/identity`'s `identityOf`, served at `GET /olai/who`), because a cell
  * would be one value for the process and this value is one value for the
  * request. The websocket cannot see it today (kolu's `serveSurfaceApp`
  * owns the upgrade).
